Add Img Cls Annotator
