+++ datadreamer/dataset_annotation/__init__.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 from .aimv2_annotator import AIMv2Annotator
 from .clip_annotator import CLIPAnnotator
+from .cls_annotator import ImgClassificationAnnotator
 from .image_annotator import BaseAnnotator, TaskList
 from .owlv2_annotator import OWLv2Annotator
 from .slimsam_annotator import SlimSAMAnnotator
@@ -11,6 +12,7 @@ __all__ = [
     "BaseAnnotator",
     "TaskList",
     "OWLv2Annotator",
+    "ImgClassificationAnnotator",
     "CLIPAnnotator",
     "SlimSAMAnnotator",
 ]
